implemented hook for repair nav
